Ensure migrations table before operating on it
Previously we had some operations that were assuming the migrations
table existed before trying to select data out of it. When the table did
not exist, you'd get an error from knex stating as such.

This patch adds code to make sure the table exists before trying to
retrieve data from it.

Closes #17.